refactor: application structure and update roadmap status (#9)
* Refactor application structure and update roadmap status

- Reorganized project structure: specific folders for Views (Editors, Layout, Explorer, Settings).
- Implemented core Services: FileOperationsService, WorkspaceService.
- Added SpreadsheetGrid control and related ViewModels.
- Completed Phases 1-6 in ROADMAP.md; Phase 7 is In Progress.
- Refined UI components and Agent interactions.

* Update app/Controls/SpreadsheetGrid.axaml.cs


* Update app/ViewModels/TableViewModel.cs


* Update app/ViewModels/TableViewModel.cs


* Update app/ViewModels/TableViewModel.cs


* Update app/ViewModels/TableViewModel.cs


* Update app/Services/FileOperationsService.cs


* Update app/Services/FileOperationsService.cs


* Update app/Services/FileOperationsService.cs


---------

+++ core/Models/JsonValidationResult.cs
@@ -12,10 +12,10 @@ public enum JsonValidationErrorType
     NullNotAllowed
 }
 
-public record JsonValidationError(int? RowIndex, string Message, string? Key, JsonValidationErrorType Type, int? LineNumber = null);
+public record JsonValidationError(int? RowIndex, string Message, string? Key, JsonValidationErrorType Type, long? LineNumber = null);
 
 public record JsonValidationResult(bool IsValid, List<JsonValidationError> Errors, JsonModel? Model)
 {
-    public static JsonValidationResult Failure(string message, JsonValidationErrorType type = JsonValidationErrorType.Syntax, int? lineNumber = null) =>
+    public static JsonValidationResult Failure(string message, JsonValidationErrorType type = JsonValidationErrorType.Syntax, long? lineNumber = null) =>
         new(false, new List<JsonValidationError> { new(null, message, null, type, lineNumber) }, null);
 }
